Rename `app` to `finances-service` so the compiled binary is well named
