+++ demo-finances-api/src/finances-app/controller/BookController.go
@@ -2,8 +2,8 @@ package controller
 
 import (
 	"github.com/gin-gonic/gin"
-	. "github.com/adeynack/learning_go/demo-finances-api/src/app/service"
-	. "github.com/adeynack/learning_go/demo-finances-api/src/app/model"
+	. "github.com/adeynack/learning_go/demo-finances-api/src/finances-app/service"
+	. "github.com/adeynack/learning_go/demo-finances-api/src/finances-app/model"
 	"net/http"
 )
