+++ demo-finances-api/src/finances-app/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	. "github.com/adeynack/learning_go/demo-finances-api/src/app/controller"
-	. "github.com/adeynack/learning_go/demo-finances-api/src/app/service"
-	. "github.com/adeynack/learning_go/demo-finances-api/src/app/repository"
+	. "github.com/adeynack/learning_go/demo-finances-api/src/finances-app/controller"
+	. "github.com/adeynack/learning_go/demo-finances-api/src/finances-app/service"
+	. "github.com/adeynack/learning_go/demo-finances-api/src/finances-app/repository"
 )
 
 func main() {
