+++ demo-finances-api/src/finances-app/model/Book.go
