+++ demo-finances-api/src/finances-app/service/AccountService.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"fmt"
 	"net/http"
-	. "github.com/adeynack/learning_go/demo-finances-api/src/app/model"
+	. "github.com/adeynack/learning_go/demo-finances-api/src/finances-app/model"
 )
 
 type AccountService struct {
